requirements.txt
Django==2.1.7
djangorestframework==3.9.1
psycopg2-binary==2.7.7
requests
dataporten-auth
python-social-auth==0.2.19
django-cors-headers
